fix(web): render all source links
