+++ web/components/chat/source-cards.tsx
@@ -1,0 +1,96 @@
+import { BookOpenText, ExternalLink } from 'lucide-react';
+
+import type { RetrievedSource } from '@/lib/api-types';
+
+import { t } from '@/lib/i18n';
+
+const SourceKindLabel = ({ source }: { source: RetrievedSource }) => (
+  <span className="rounded-full border border-border/70 px-1.5 py-0.5 text-[10px] font-medium uppercase tracking-wide text-muted-foreground">
+    {source.kind === 'faq' ? t('sources.faq') : t('sources.chunk')}
+  </span>
+);
+
+const SourceCard = ({ source }: { source: RetrievedSource }) => {
+  const links = source.links ?? [];
+  const title = source.section
+    ? `${source.title} · ${source.section}`
+    : source.title;
+
+  return (
+    <li className="min-w-0 rounded-lg border border-border/70 bg-muted/20 p-3 transition-colors hover:bg-muted/35">
+      <div className="flex items-start justify-between gap-3">
+        <div className="min-w-0 space-y-1">
+          <div className="flex flex-wrap items-center gap-2">
+            <SourceKindLabel source={source} />
+            {typeof source.chunkIndex === 'number' ? (
+              <span className="text-[10px] text-muted-foreground/70">
+                #{source.chunkIndex + 1}
+              </span>
+            ) : null}
+          </div>
+          <p className="line-clamp-2 text-sm font-medium leading-snug text-foreground">
+            {title}
+          </p>
+        </div>
+        {links.length > 0 ? (
+          <div className="flex shrink-0 flex-wrap justify-end gap-1">
+            {links.map((link) => (
+              <a
+                aria-label={`${t('sources.link')}: ${link.label}`}
+                className="rounded-md p-1 text-muted-foreground hover:bg-background hover:text-foreground"
+                href={link.url}
+                key={`${link.label}:${link.url}`}
+                rel="noreferrer"
+                target="_blank"
+              >
+                <ExternalLink
+                  aria-hidden="true"
+                  className="size-3.5"
+                />
+              </a>
+            ))}
+          </div>
+        ) : null}
+      </div>
+      {source.snippet ? (
+        <p className="mt-2 line-clamp-2 text-xs leading-relaxed text-muted-foreground">
+          {source.snippet}
+        </p>
+      ) : null}
+    </li>
+  );
+};
+
+export const SourceCards = ({
+  sources,
+}: {
+  sources: readonly RetrievedSource[];
+}) => {
+  if (sources.length === 0) {
+    return null;
+  }
+
+  return (
+    <section
+      aria-label={t('sources.title')}
+      className="mt-3 space-y-2"
+      data-testid="message-sources"
+    >
+      <div className="flex items-center gap-1.5 text-xs font-medium text-muted-foreground">
+        <BookOpenText
+          aria-hidden="true"
+          className="size-3.5"
+        />
+        <span>{t('sources.title')}</span>
+      </div>
+      <ul className="grid gap-2 sm:grid-cols-2">
+        {sources.map((source) => (
+          <SourceCard
+            key={`${source.kind}:${source.id}`}
+            source={source}
+          />
+        ))}
+      </ul>
+    </section>
+  );
+};
